layout: rename manager module to lifecycle_manager
Rename the lifecycle manager module as suggested by @cjp256.
