+++ craft_parts/__init__.py
@@ -20,6 +20,6 @@ __version__ = "0.0.1"  # noqa: F401
 
 from .actions import Action, ActionType  # noqa: F401
 from .infos import ProjectInfo  # noqa: F401
-from .manager import LifecycleManager  # noqa: F401
+from .lifecycle_manager import LifecycleManager  # noqa: F401
 from .parts import Part  # noqa: F401
 from .steps import Step  # noqa: F401
